Fix #4421: Implements Rattattack Battle Mystery Encounter
Adds a new Mystery Encounter,
a Rattata/ Raticate battle, which is inspired by the games.
This battle only appears, between in waves 31 to 179.
Co-authored-by: sophiaaa05 <sophia.alencar@tecnico.ulisboa.pt>

Fix #4421: Implements
Rattattack Battle Mystery
Encounter

Adds a new Mystery Encounter,
a Rattata/ Raticate battle, which is inspired by the games.
This battle only appears, between in waves 31 to 179.
Co-authored-by: sophiaaa05 <sophia.alencar@tecnico.ulisboa.pt>